Project structure fix
