+++ habr/templates/base.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+    <meta charset="UTF-8">
+    <link href="{{ url_for('static', filename='css/bootstrap.css') }}" rel="stylesheet">
+    <title>{% block title %}{% endblock %}</title>
+</head>
+<body>
+    {% if title -%}
+    <h1>{{title}}</h1>
+    {% else -%}
+    <h1>ХАБР</h1>
+    {% endif %}
+    {% block content %}{% endblock %}
+    <script src="{{ url_for('static', filename='js/bootstrap.bundle.js') }}"></script>
+</script>
+</body>
+</html>
