+++ habr/templates/index.html
@@ -1,6 +1,5 @@
 {% extends 'base.html' %}
 
 {% block content %}
-{{ super() }}
-<p>Страница "О сайте"
+<h1 style="color: blue;">Hello! This is main page.</h1>
 {% endblock %}
