+++ run.bat
@@ -1,0 +1,1 @@
+flask run
