--- templates/base.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    {% block title -%}
-    {% if title %}
-    <title>Хабр - {{title}}</title>
-    {% else %}
-    <title>Хабр</title>
-    {% endif %}
-    {% endblock %}
-</head>
-<body>
-    {% block content -%}
-    {% block mainmenu -%}
-    <ul>
-        {% for m in menu %}
-        <li> {{m}}</li>
-        {% endfor %}
-    </ul>
-    {% endblock mainmenu -%}
-
-    {% if title -%}
-    <h1>{{title}}</h1>
-    {% else -%}
-    <h1>ХАБР</h1>
-    {% endif %}
-    {% endblock -%}
-    
-</body>
-</html>
