--- templates/index.html
@@ -1,6 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-{{ super() }}
-<p>Главная страница
-{% endblock %}
