+++ wsgi.py
@@ -1,0 +1,3 @@
+from habr.app import create_app
+
+app = create_app()
